exported arrayFormatter to helper util function.
